NEW Added a Tab View to the App
Added a Tab View to the app.

NOTE - created the tab view in ContentView() in order to not have to touch Scene Delegate. If you want the ContentView to be a page within the Tab View, make sure to set Scene Delegate start to the Tab View and not Content View.